STCOM-855 publish storybook doc via GitHub pages
Replace the existing `./docs` directory with storybook output. Note that
`build-storybook` includes a "clean" directive that removes the output
directory, meaning it is a requirement to specify an output directory
because if you specify `.` then `build-storybook` will attempt to remove
the current working directory, and that surely is not what you want.

Refs STCOM-855